Scrolling Coordinator must be deleted on the main thread https://bugs.webkit.org/show_bug.cgi?id=78756 <rdar://problem/10866167>
Reviewed by Sam Weinig.

ScrollingTree::invalidate will finish breaking the cycle between the scrolling coordinator and the
scrolling tree by dereffing the scrolling coordinator. We need to make sure that this happens on the main
thread because the scrolling coordinator expects to be destroyed from there.

* page/scrolling/ScrollingTree.cpp:
(WebCore::derefScrollingCoordinator):
(WebCore):
(WebCore::ScrollingTree::invalidate):

Identifier: 95688@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@107849 268f45cc-cd09-0410-ab3c-d52691b4dbfc