Build fix for Qt 5 without QtWidgets.
Reviewed by Tor Arne Vestbø.

The reason why QGLContext is included in Extensions3DQt.cpp is to
achieve an implicit gl.h inclusion (needed for GL_FALSE). This patch
replaces the inclusion with the OpenGLShims.h inclusions, which has
the necessary #ifdefs in place to pull in gl.h with Qt 5 without QtWidgets.

* platform/graphics/qt/Extensions3DQt.cpp:

Identifier: 95758@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@107922 268f45cc-cd09-0410-ab3c-d52691b4dbfc