webkitpy: fix test code after bug 78181 https://bugs.webkit.org/show_bug.cgi?id=78870
Reviewed by Adam Barth.

I forgot to update the test code as per Tony's suggestion in
bug 78181 (r108005). This change does that, and in doing so
it revealed a subtle bug in _InlineWorkerConnection.run() that
I am fixing as part of this (_alive would not be cleared if an
exception was thrown).

* Scripts/webkitpy/layout_tests/controllers/manager_worker_broker.py:
(_InlineWorkerConnection.run):
* Scripts/webkitpy/layout_tests/controllers/manager_worker_broker_unittest.py:
(_TestWorker.__init__):
(_TestWorker.handle_stop):
(_TestWorker.handle_test):
(_TestWorker.run):
(_TestsMixin.handle_exception):
(_TestsMixin.test_unknown_message):

Identifier: 95971@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@108141 268f45cc-cd09-0410-ab3c-d52691b4dbfc