webkitpy: speed up FileLockTest.test_lock_lifecycle https://bugs.webkit.org/show_bug.cgi?id=79189
Reviewed by Adam Barth.

The unit test had a second lock trying to acquire a
lock already held, and a hard-coded timeout of one second.
I have changed the timeouts to a tenth of a second (which
is probably still fine) and changed the hard-spin on
the lock to a slow-spin every hundredth of a second.

* Scripts/webkitpy/common/system/file_lock.py:
(FileLock.acquire_lock):
* Scripts/webkitpy/common/system/file_lock_integrationtest.py:
(FileLockTest.setUp):
(FileLockTest.test_stuck_lock):

Identifier: 96313@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@108514 268f45cc-cd09-0410-ab3c-d52691b4dbfc