[chromium] Unreviewed gardening. Apply bug number to svg/carto.net failures. https://bugs.webkit.org/show_bug.cgi?id=79423
These two tests were in a set of tests that needed to be rebaselined
after a jpeg decoder change. However, there was also a text change
during this period that caused debug vs. release differences for text
spans in the SVG output. Filing a real bug so this can get looked at.

* platform/chromium/test_expectations.txt:

Identifier: 96482@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@108704 268f45cc-cd09-0410-ab3c-d52691b4dbfc