getComputedStyle fails for 'first-line' pseudo-element https://bugs.webkit.org/show_bug.cgi?id=57505
Reviewed by Tony Chang.

Source/WebCore:

Querying the selector with a pseudo-element using getComputedStyle should work
even if the selector was not declared in the stylesheet. When not declared, we need
to use the RenderStyle created to do the rendering as there is no pseudo-style.
This match the behavior of Firefox.

No new tests : Updated expectation and extended getComputedStyle-with-pseudo-element.html.

* dom/Element.cpp:
(WebCore::Element::computedStyle):

LayoutTests:

Update expected results to the new behavior. Also added some extra checks on
other pseudo-elements.

* fast/css/getComputedStyle/getComputedStyle-with-pseudo-element-expected.txt:
* fast/css/getComputedStyle/getComputedStyle-with-pseudo-element.html:


Identifier: 96894@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@109134 268f45cc-cd09-0410-ab3c-d52691b4dbfc