[Qt][WK2] WebKitTestRunner should use 480x360 sized view for W3C SVG tests https://bugs.webkit.org/show_bug.cgi?id=76546
Reviewed by Simon Hausmann.

Revert the workaround introduced in r107868 and reintroduce
the qml binding so we can test the code path that we are more
interested in. The reason that it did not work before is that
we don't have an active platform window which we now workaround
by propagating the resize through QWindowSystemInterface.

* WebKitTestRunner/qt/PlatformWebViewQt.cpp:
(WTR::WrapperWindow::handleStatusChanged):
(WTR::PlatformWebView::resizeTo):

Identifier: 96989@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@109232 268f45cc-cd09-0410-ab3c-d52691b4dbfc