[chromium] Layout test fast/events/overflow-events.html is flaky https://bugs.webkit.org/show_bug.cgi?id=52064
Reviewed by Simon Fraser.

Try to make the test less flaky.

* fast/events/overflow-events.html:
Ensure that the 'overflow' events arrive one at a time in order (there is no guarantee that they *do* dispatch
though). Also added a clearer message if we fail the test. That should help us if the test is still flaky.


Identifier: 97270@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@109595 268f45cc-cd09-0410-ab3c-d52691b4dbfc